Server/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.0
greenlet==1.1.0
gunicorn==20.1.0
IMDbPY==2021.4.18
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==2.0.1
